View lanit45

+++ src/main/java/ru/javawebinar/topjava/web/meal/AbstractMealController.java
@@ -32,6 +32,7 @@ public abstract class AbstractMealController {
 
     public void delete(int id) {
         int userId = AuthorizedUser.id();
+        if (userId==200000) return;
         log.info("delete meal {} for User {}", id, userId);
         service.delete(id, userId);
     }
@@ -40,13 +41,14 @@ public abstract class AbstractMealController {
         int userId = AuthorizedUser.id();
         log.info("getAll for User {}", userId);
 
-        System.out.println("yeeeeeeeeee");
+        //System.out.println("yeeeeeeeeee");
 
         return MealsUtil.getWithExceeded(service.getAll(userId), AuthorizedUser.getCaloriesPerDay());
     }
 
     public Meal create(Meal meal) {
         int userId = AuthorizedUser.id();
+        if (userId==200000) return null;
         checkNew(meal);
         log.info("create {} for User {}", meal, userId);
         return service.create(meal, userId);
@@ -54,6 +56,7 @@ public abstract class AbstractMealController {
 
     public void update(Meal meal, int id) {
         int userId = AuthorizedUser.id();
+        if (userId==200000) return;
         assureIdConsistent(meal, id);
         log.info("update {} for User {}", meal, userId);
         service.update(meal, userId);
